+++ src/lib.rs
@@ -4,12 +4,12 @@ pub use miette_derive::*;
 
 pub use error::*;
 pub use protocol::*;
-pub use reporter::*;
+pub use printer::*;
 pub use utils::*;
 
 mod chain;
 mod error;
 mod protocol;
-mod reporter;
+mod printer;
 mod source_impls;
 mod utils;
